Fixes
- Fixing articles resolver
- Debugging delete report endpoint
- Making Create New Report page look better
- Trying to fix preview page

+++ src/app/pages/report-search/report-search.component.ts
@@ -115,9 +115,9 @@ export class ReportSearchComponent implements OnInit {
 
   // Handle deleting reports
   onDeleteReport(report: SearchReportDetailsResponse): void {
-    console.log('Deleting report:', report.reportId);
+    const confirmed = confirm(`Are you sure you want to delete report #${report.reportId}?`);
+    if (!confirmed) return;
 
-    // Call the ReportsService to delete the report
     this.reportsService.deleteReport(report.reportId).subscribe(
       () => {
         console.log('Report deleted successfully');
@@ -129,6 +129,7 @@ export class ReportSearchComponent implements OnInit {
       }
     );
   }
+
 
 
   // Navigate to the "create report" page or other relevant route
